Extract selector into a hook
